split test file into test dir
